+++ benchmarks/Unary/SignMatrixBench.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Tensor\Benchmarks\Unary;
+
+use Tensor\Matrix;
+
+/**
+ * @Groups({"Unary"})
+ * @BeforeMethods({"setUp"})
+ */
+class SignMatrixBench
+{
+    /**
+     * @var Matrix
+     */
+    protected $a;
+
+    public function setUp() : void
+    {
+        $this->a = Matrix::uniform(1000, 1000);
+    }
+
+    /**
+     * @Subject
+     * @Iterations(5)
+     * @OutputTimeUnit("seconds", precision=3)
+     */
+    public function sign() : void
+    {
+        $this->a->sign();
+    }
+}